(#87) Update to Cake 2.0.0
The implementation of ICakeArguments had to be extended due to the
addition of a new GetArguments() method.

The targets for the projects have been updated to remove .NET 4.6.1.

The references for Cake have been updated to use 2.0.0, the only
exception to this is within the Tests project, where we can make use of
the latest version.